pyproject.toml
[project]
name = "fastapi-auth-gateway"
version = "0.1.0"
description = ""
authors = [
    {name = "ankitraj20616",email = "ankitrajthakur803@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "supabase (>=2.27.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "fastapi_auth_gateway", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
